pyproject.toml
[build-system]
requires = ["setuptools>=45", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "semantic"
version = "0.1.0"
description = "A CLI tool that generates semantic summaries of codebases for AI coding agents"
readme = "README.md"
requires-python = ">=3.10"
authors = [
    {name = "Semantic Team"}
]
keywords = ["cli", "codebase", "analysis", "ai", "summarization"]
classifiers = [
    "Development Status :: 3 - Alpha",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
]

dependencies = [
    "pydantic>=2.0.0",
    "typer[all]>=0.9.0",
    "pathspec>=0.11.0",
    "python-dateutil>=2.8.0",
    "openai>=1.0.0",
    "PyYAML>=6.0.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.0.0",
    "pytest-cov>=4.0.0",
    "black>=23.0.0",
    "isort>=5.12.0",
    "flake8>=6.0.0",
    "mypy>=1.0.0",
]
llm = [
    "anthropic>=0.23.0",
    "google-generativeai>=0.3.0",
]

[project.scripts]
semantic = "codebase_summarizer.main:app"

[tool.setuptools.packages.find]
where = ["src"]

[tool.setuptools.package-dir]
"" = "src"

[tool.black]
line-length = 88
target-version = ['py310']

[tool.isort]
profile = "black"
src_paths = ["src", "tests"]

[tool.mypy]
python_version = "3.10"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = true